requirements/dev.txt
-r base.txt
ipython==8.25.0
psycopg-binary==3.2.1
django-multiselectfield==0.1.13
pillow==10.4.0
django-extensions==3.2.3
Faker==27.0.0
